Add endpoint for WeitereZutaten

+++ src/main/java/com/sqlite/demo/model/lager/Hefe.java
@@ -17,6 +17,6 @@ public class Hefe extends Lagerbestand {
     private long id;
 
     private String name = super.getName();
-    private int menge = super.getMenge();
+    private float menge = super.getMenge();
     private float preis = super.getPreis();
 }
